upgrading to selenium for source code scraping
this allowed for a more universal retrieval of data.  for celebheights and celebritynetworths, they sometimes had different formats across pages, which caused errors when I tried to retrieve specific data.  So although selenium is a lot slower, it allows for a flawless retrieval of data, because it allows for the Javascript to load.