Implementing the Fundamentals
This commit provides essentially everything needed on the web end of this project.  It allows for the importing of the Data JSON and subsequently, the dynamic creation of listings.  It also includes several other methods (and CSS) that help the website come together nicely.  There are still minor updates I plan to add but this is the final large commit to this project.